requirements.txt
# Python 3.6.6

absl-py==0.5.0
astor==0.7.1
cycler==0.10.0
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
kiwisolver==1.0.1
Markdown==3.0
matplotlib==3.0.0
numpy==1.14.5
protobuf==3.6.1
pyparsing==2.2.1
python-dateutil==2.7.3
six==1.11.0
tensorboard==1.10.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
Werkzeug==0.14.1
